moved methods from AttachedFileController to TalkController. Deleted AttachedFileControllerTest
