+++ webapp/src/test/java/ua/rd/cm/web/controller/TalkControllerTest.java
@@ -26,11 +26,10 @@ import ua.rd.cm.config.TestSecurityConfig;
 import ua.rd.cm.config.WebMvcConfig;
 import ua.rd.cm.config.WebTestConfig;
 import ua.rd.cm.domain.*;
+import ua.rd.cm.dto.MessageDto;
 import ua.rd.cm.services.TalkService;
 import ua.rd.cm.services.UserInfoService;
 import ua.rd.cm.services.UserService;
-import ua.rd.cm.services.exception.TalkNotFoundException;
-import ua.rd.cm.dto.MessageDto;
 import ua.rd.cm.dto.TalkDto;
 
 import javax.servlet.Filter;
